+++ tfjs-core/src/gradients/Max_grad.ts
@@ -18,7 +18,7 @@
 import {Max, MaxAttrs} from '../kernel_names';
 import {GradConfig, NamedAttrMap} from '../kernel_registry';
 import * as axis_util from '../ops/axis_util';
-import {gradForMinAndMax} from '../ops/reduction_ops';
+import {gradForMinAndMax} from '../ops/reduction_ops_util';
 import {transpose} from '../ops/transpose';
 import {Tensor} from '../tensor';
 import * as util from '../util';
